services/tts/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
piper-tts==1.2.0
numpy==1.24.3
scipy==1.11.4
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2

